@@ -2,6 +2,7 @@
 #include <array>
 #include <cstdint>
 #include <filesystem>
+#include <format>
 #include <fstream>
 #include <iostream>
 #include <streambuf>
@@ -17,8 +18,8 @@
 
 #include <GLFW/glfw3.h>
 
-constexpr int windowStartWidth = 1280;
-constexpr int windowStartHeight = 720;
+constexpr int window_start_width = 1280;
+constexpr int window_start_height = 720;
 
 template <typename T>
 int opengl_test(std::vector<cl_float4> &data, T &&callable) {
@@ -31,13 +32,13 @@ int opengl_test(std::vector<cl_float4> &data, T &&callable) {
         return 1;
     }
 
-    glfwSetErrorCallback([](int errCode, const char *msg) {
-        spdlog::error("glfw errored with code {}: {}", errCode, msg);
+    glfwSetErrorCallback([](int err_code, const char *msg) {
+        spdlog::error("glfw errored with code {}: {}", err_code, msg);
     });
 
     // Create window with graphics context
-    GLFWwindow *window = glfwCreateWindow(windowStartWidth, windowStartHeight,
-                                          "test", nullptr, nullptr);
+    GLFWwindow *window = glfwCreateWindow(
+        window_start_width, window_start_height, "test", nullptr, nullptr);
     if (window == nullptr) {
         spdlog::error("glfwCreateWindow failed");
         return 1;
@@ -49,8 +50,8 @@ int opengl_test(std::vector<cl_float4> &data, T &&callable) {
         return 1;
     }
 
-    int screen_width;
-    int screen_height;
+    int screen_width{};
+    int screen_height{};
     glfwGetFramebufferSize(window, &screen_width, &screen_height);
     glViewport(0, 0, screen_width, screen_height);
 
@@ -64,14 +65,14 @@ int opengl_test(std::vector<cl_float4> &data, T &&callable) {
     // Setup Dear ImGui style
     ImGui::StyleColorsDark();
 
-    constexpr int textureWidth = 512;
-    constexpr int textureHeight = 512;
+    constexpr int texture_width = 512;
+    constexpr int texture_height = 512;
 
     glDebugMessageCallback(
-        [](GLenum source, GLenum type, GLuint id, GLenum severity,
+        [](GLenum source, GLenum type, GLuint id, GLenum severity, // NOLINT
            [[maybe_unused]] GLsizei length, const GLchar *message,
-           [[maybe_unused]] const void *userParam) {
-            auto formated = fmt::format("opengl debug: source: {} type: {} id: "
+           [[maybe_unused]] const void *user_param) {
+            auto formated = std::format("opengl debug: source: {} type: {} id: "
                                         "{} severity: {}\n message: {}",
                                         source, type, id, severity, message);
             switch (severity) {
@@ -87,21 +88,24 @@ int opengl_test(std::vector<cl_float4> &data, T &&callable) {
             case GL_DEBUG_SEVERITY_HIGH:
                 spdlog::error(formated);
                 break;
+            default:
+                spdlog::error("Unknown open error code {}", formated);
+                break;
             }
         },
         nullptr);
     glEnable(GL_DEBUG_OUTPUT);
 
-    GLuint texture;
+    GLuint texture{};
     glGenTextures(1, &texture);
     glBindTexture(GL_TEXTURE_2D, texture);
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB32F, textureWidth, textureHeight, 0,
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB32F, texture_width, texture_height, 0,
                  GL_RGBA, GL_FLOAT, data.data());
     glBindTexture(GL_TEXTURE_2D, 0);
 
-    GLuint readFboId = 0;
-    glGenFramebuffers(1, &readFboId);
-    glBindFramebuffer(GL_READ_FRAMEBUFFER, readFboId);
+    GLuint fbo_id = 0;
+    glGenFramebuffers(1, &fbo_id);
+    glBindFramebuffer(GL_READ_FRAMEBUFFER, fbo_id);
     glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_COLOR_ATTACHMENT0,
                            GL_TEXTURE_2D, texture, 0);
     glBindFramebuffer(GL_READ_FRAMEBUFFER, 0);
@@ -109,21 +113,21 @@ int opengl_test(std::vector<cl_float4> &data, T &&callable) {
     while (glfwWindowShouldClose(window) != GLFW_TRUE) {
         glfwPollEvents();
 
-        int display_w;
-        int display_h;
+        int display_w{};
+        int display_h{};
         glfwGetFramebufferSize(window, &display_w, &display_h);
 
         glClearColor(0, 0, 0, 1);
         glClear(GL_COLOR_BUFFER_BIT);
 
         // if Changed
-        glTextureSubImage2D(texture, 0, 0, 0, textureWidth, textureHeight,
+        glTextureSubImage2D(texture, 0, 0, 0, texture_width, texture_height,
                             GL_RGBA, GL_FLOAT, data.data());
-        glBindFramebuffer(GL_READ_FRAMEBUFFER, readFboId);
-        glBlitFramebuffer(0, 0, std::min(textureWidth, display_w),
-                          std::min(textureHeight, display_h), 0,
-                          display_h - std::min(textureHeight, display_h),
-                          std::min(textureWidth, display_w), display_h,
+        glBindFramebuffer(GL_READ_FRAMEBUFFER, fbo_id);
+        glBlitFramebuffer(0, 0, std::min(texture_width, display_w),
+                          std::min(texture_height, display_h), 0,
+                          display_h - std::min(texture_height, display_h),
+                          std::min(texture_width, display_w), display_h,
                           GL_COLOR_BUFFER_BIT, GL_LINEAR);
         glBindFramebuffer(GL_READ_FRAMEBUFFER, 0);
 
@@ -134,7 +138,8 @@ int opengl_test(std::vector<cl_float4> &data, T &&callable) {
 
         // render your GUI
         ImGui::Begin("Demo window");
-        ImGui::Text("framerate %.2f", io.Framerate);
+        std::string formated = std::format("framerate {:.3f}", io.Framerate);
+        ImGui::TextUnformatted(formated.c_str());
         callable();
         ImGui::End();
 
