@@ -1,30 +1,35 @@
 
-#ifdef __OPENCL_C_VERSION__
-#define double3_vec double3
-#else
+#ifdef __cplusplus
 #include <CL/cl2.hpp>
-#define double3_vec cl_double3
+extern "C" {
+#define double3_vec cl_double3 // NOLINT
+#else
+#define double3_vec double3
 #endif
 
-struct ray {
-    double3_vec origin;
-    double3_vec direction;
+struct Ray {
+    double3_vec m_origin;
+    double3_vec m_direction;
 };
 
-struct hit_record {
-    double3_vec p;
-    double3_vec normal;
-    double t;
+struct HitRecord {
+    double3_vec m_p;
+    double3_vec m_normal;
+    double m_t;
 };
 
-struct circle {
-    double3_vec center;
-    double radius;
+struct Circle {
+    double3_vec m_center;
+    double m_radius;
 };
 
-struct camera {
-    double3_vec origin;
-    double3_vec lower_left_corner;
-    double3_vec horizontal;
-    double3_vec vertical;
+struct Camera {
+    double3_vec m_origin;
+    double3_vec m_lower_left_corner;
+    double3_vec m_horizontal;
+    double3_vec m_vertical;
 };
+
+#ifdef __cplusplus
+}
+#endif
